requirements.txt
pandas
numpy
scikit-learn
xgboost
openpyxl
mysql-connector-python
python-dotenv
streamlit
